Add Pagination scroll listener.
In this commit, I add a class that handles the pagination through the recyclerView "PaginationScrollListener".
I also add a new function to add the new data to the images list "ImagesAdapter".
Functions trending and search from the API are modified. Add offset, it will be used for pagination. I also
modified the function default parameters.
In the TrendingFragment, I add another function the fetch more data and add it to the imagesList.

by m68476521.1@gmail.com